pyproject.toml
[project]
name = "panager"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.4.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain-openai>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "apscheduler>=3.10.0",
    "sentence-transformers>=3.3.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.154.0",
    "psycopg[binary]>=3.2.0",
    "torch",
    "notion-client>=3.0.0",
    "httpx>=0.28.1",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/panager"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "watchfiles>=1.0.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
norecursedirs = ["legacy"]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
